Extract common functionality into property version hook
